Add a fast-cap-dirty PTE behavior
The cap-store instructions now look at the data being stored to decide
whether they issue Write(Cap) or Write(Data) requests on "the bus".
This allows the PTW logic (and update_PTE_Bits) in particular to not
fast-cap-dirty a page that's being targeted by the store of an untagged
capability.  This is not, however, viable for AMOCAS and so additional
changes may be required if we are to avoid considering AMOCAS as always
capability-dirtying.